+++ packages/core/src/build-module.ts
@@ -1,5 +1,5 @@
-import { IgnitionError } from "./errors";
-import { ERRORS } from "./errors-list";
+import { IgnitionError } from "./internal/errors";
+import { ERRORS } from "./internal/errors-list";
 import { ModuleConstructor } from "./internal/module-builder";
 import { isValidIgnitionIdentifier } from "./internal/utils/identifier-validators";
 import { IgnitionModule, IgnitionModuleResult } from "./types/module";
