+++ packages/core/src/internal/errors.ts
