+++ packages/core/src/internal/execution/jsonrpc-client.ts
@@ -1,6 +1,6 @@
-import { IgnitionError } from "../../errors";
-import { ERRORS } from "../../errors-list";
 import { EIP1193Provider } from "../../types/provider";
+import { IgnitionError } from "../errors";
+import { ERRORS } from "../errors-list";
 
 import {
   NetworkFees,
