+++ packages/core/src/internal/execution/libraries.ts
@@ -13,9 +13,9 @@
  * @file
  */
 
-import { IgnitionError } from "../../errors";
-import { ERRORS } from "../../errors-list";
 import { Artifact } from "../../types/artifact";
+import { IgnitionError } from "../errors";
+import { ERRORS } from "../errors-list";
 import { assertIgnitionInvariant } from "../utils/assertions";
 
 /**
