+++ packages/core/src/internal/execution/nonce-management/get-nonce-sync-messages.ts
@@ -1,7 +1,5 @@
 import uniq from "lodash/uniq";
 
-import { IgnitionError } from "../../../errors";
-import { ERRORS } from "../../../errors-list";
 import {
   isArtifactContractAtFuture,
   isNamedContractAtFuture,
@@ -12,6 +10,8 @@ import {
   IgnitionModule,
   IgnitionModuleResult,
 } from "../../../types/module";
+import { IgnitionError } from "../../errors";
+import { ERRORS } from "../../errors-list";
 import { getFuturesFromModule } from "../../utils/get-futures-from-module";
 import { getPendingOnchainInteraction } from "../../views/execution-state/get-pending-onchain-interaction";
 import { resolveFutureFrom } from "../future-processor/helpers/future-resolvers";
