+++ packages/core/src/internal/execution/nonce-management/json-rpc-nonce-manager.ts
@@ -1,5 +1,5 @@
-import { IgnitionError } from "../../../errors";
-import { ERRORS } from "../../../errors-list";
+import { IgnitionError } from "../../errors";
+import { ERRORS } from "../../errors-list";
 import { JsonRpcClient } from "../jsonrpc-client";
 
 /**
