+++ packages/core/src/internal/module-builder.ts
@@ -1,7 +1,5 @@
 import { inspect } from "util";
 
-import { IgnitionError } from "../errors";
-import { ERRORS } from "../errors-list";
 import {
   isAccountRuntimeValue,
   isAddressResolvableFuture,
@@ -48,6 +46,8 @@ import {
   StaticCallOptions,
 } from "../types/module-builder";
 
+import { IgnitionError } from "./errors";
+import { ERRORS } from "./errors-list";
 import { equalAddresses, isAddress } from "./execution/utils/address";
 import {
   AccountRuntimeValueImplementation,
