+++ packages/core/src/internal/utils/assertions.ts
@@ -1,5 +1,5 @@
-import { IgnitionError } from "../../errors";
-import { ERRORS } from "../../errors-list";
+import { IgnitionError } from "../errors";
+import { ERRORS } from "../errors-list";
 
 export function assertIgnitionInvariant(
   invariant: boolean,
