+++ packages/core/src/internal/validation/futures/validateArtifactContractDeployment.ts
@@ -1,5 +1,3 @@
-import { IgnitionError } from "../../../errors";
-import { ERRORS } from "../../../errors-list";
 import {
   isAccountRuntimeValue,
   isModuleParameterRuntimeValue,
@@ -7,6 +5,8 @@ import {
 import { ArtifactResolver } from "../../../types/artifact";
 import { DeploymentParameters } from "../../../types/deploy";
 import { ContractDeploymentFuture } from "../../../types/module";
+import { IgnitionError } from "../../errors";
+import { ERRORS } from "../../errors-list";
 import { validateContractConstructorArgsLength } from "../../execution/abi";
 import { validateLibraryNames } from "../../execution/libraries";
 import {
