+++ packages/core/src/internal/validation/futures/validateNamedContractAt.ts
@@ -1,5 +1,3 @@
-import { IgnitionError } from "../../../errors";
-import { ERRORS } from "../../../errors-list";
 import {
   isArtifactType,
   isModuleParameterRuntimeValue,
@@ -7,6 +5,8 @@ import {
 import { ArtifactResolver } from "../../../types/artifact";
 import { DeploymentParameters } from "../../../types/deploy";
 import { NamedArtifactContractAtFuture } from "../../../types/module";
+import { IgnitionError } from "../../errors";
+import { ERRORS } from "../../errors-list";
 
 export async function validateNamedContractAt(
   future: NamedArtifactContractAtFuture<string>,
