+++ packages/core/src/internal/validation/futures/validateNamedStaticCall.ts
@@ -1,5 +1,3 @@
-import { IgnitionError } from "../../../errors";
-import { ERRORS } from "../../../errors-list";
 import {
   isAccountRuntimeValue,
   isArtifactType,
@@ -8,6 +6,8 @@ import {
 import { ArtifactResolver } from "../../../types/artifact";
 import { DeploymentParameters } from "../../../types/deploy";
 import { StaticCallFuture } from "../../../types/module";
+import { IgnitionError } from "../../errors";
+import { ERRORS } from "../../errors-list";
 import {
   validateArtifactFunction,
   validateFunctionArgumentParamType,
