+++ packages/core/src/internal/validation/futures/validateSendData.ts
@@ -1,5 +1,3 @@
-import { IgnitionError } from "../../../errors";
-import { ERRORS } from "../../../errors-list";
 import {
   isAccountRuntimeValue,
   isModuleParameterRuntimeValue,
@@ -7,6 +5,8 @@ import {
 import { ArtifactResolver } from "../../../types/artifact";
 import { DeploymentParameters } from "../../../types/deploy";
 import { SendDataFuture } from "../../../types/module";
+import { IgnitionError } from "../../errors";
+import { ERRORS } from "../../errors-list";
 import { validateAccountRuntimeValue } from "../utils";
 
 export async function validateSendData(
