+++ packages/core/src/internal/wiper.ts
@@ -1,7 +1,6 @@
-import { IgnitionError } from "../errors";
-import { ERRORS } from "../errors-list";
-
 import { DeploymentLoader } from "./deployment-loader/types";
+import { IgnitionError } from "./errors";
+import { ERRORS } from "./errors-list";
 import {
   applyNewMessage,
   loadDeploymentState,
